replica update test
