+++ integration_tests/src/tests/mod.rs
@@ -24,7 +24,8 @@ mod power_scenario_tests;
 pub use power_scenario_tests::*;
 mod publish_deals_test;
 pub use publish_deals_test::*;
-// mod replica_update_test;
+mod replica_update_test;
+pub use replica_update_test::*;
 // mod terminate_test;
 // mod verified_claim_test;
 // mod verifreg_remove_datacap_test;
